create separate file for visualize view controller tests
